# server/docs/swagger.yaml
basePath: /v1
definitions:
  domain.AddUser:
    properties:
      name:
        example: Luther
        type: string
      password:
        example: umbrella
        type: string
    type: object
  domain.Chatroom:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      limit:
        type: integer
      name:
        type: string
      updatedAt:
        type: string
      users:
        items:
          $ref: '#/definitions/domain.User'
        type: array
    type: object
  domain.ResponseUser:
    properties:
      id:
        type: integer
      name:
        type: string
      token:
        type: string
    type: object
  domain.User:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      updatedAt:
        type: string
    type: object
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  rest.ChatroomUrlResponse:
    properties:
      url:
        type: string
    type: object
  rest.CreateChatroomDTO:
    properties:
      name:
        example: new chatroom
        type: string
    type: object
  rest.ResponseError:
    properties:
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: dnay2k@gmail.com
    name: 담당자 마운틴
  description: 소켓을 이용한 채팅서버 API 문서
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Chatserver Tutorial API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
      - application/json
      description: login with id and password
      parameters:
      - description: name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.AddUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer XXX
              type: string
          schema:
            $ref: '#/definitions/domain.ResponseUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Login
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: create user with name and password
      parameters:
      - description: name
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.AddUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Register
      tags:
      - auth
  /rooms:
    get:
      consumes:
      - application/json
      description: Get list of chatrooms
      parameters:
      - description: Bearer XXX
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/domain.Chatroom'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Fetch Chatrooms
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Create new chatroom
      parameters:
      - description: Bearer XXX
        in: header
        name: Authorization
        required: true
        type: string
      - description: create chatroom dto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/rest.CreateChatroomDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Chatroom'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Create chatroom
      tags:
      - chat
  /rooms/{roomId}:
    delete:
      consumes:
      - application/json
      description: Delete chatroom
      parameters:
      - description: Bearer XXX
        in: header
        name: Authorization
        required: true
        type: string
      - description: delete chatroom id
        in: path
        name: roomId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Chatroom'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Delete chatroom
      tags:
      - chat
  /rooms/{roomId}/open:
    get:
      consumes:
      - application/json
      description: get chatroom socket
      parameters:
      - description: Bearer XXX
        in: header
        name: Authorization
        required: true
        type: string
      - description: chatroom id
        in: path
        name: roomId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/rest.ChatroomUrlResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/rest.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/rest.ResponseError'
      summary: Find chatroom socket address
      tags:
      - chat
swagger: "2.0"